Moving testSpec to correct place and some minor changes
